requirements.txt
install==1.3.5
pika==1.3.2
python-dotenv==1.0.0
